Adding Apply test for Transactions
