+++ testing/cli-e2etest/testscenarios/transaction/resources/new-transaction.yaml
@@ -4,5 +4,5 @@ spec:
   name: New Transaction
   description: a transaction
   steps:
-  - ./transaction-step-1.yaml
-  - ./transaction-step-2.yaml
+  - 9wtAH2_Vg
+  - ajksdkasjbd
