+++ testing/cli-e2etest/testscenarios/transaction/resources/updated-new-transaction.yaml
@@ -1,0 +1,9 @@
+type: Transaction
+spec:
+  id: Qti5R3_VR
+  name: Updated Transaction
+  description: an updated transaction
+  steps:
+  - 9wtAH2_Vg
+  - ajksdkasjbd
+  - ajksdkasjbd
